Move version to specific package
