+++ nomad-exporter.go
@@ -13,6 +13,7 @@ import (
 	"time"
 
 	"github.com/hashicorp/nomad/api"
+	"github.com/pcarranza/nomad-exporter/version"
 	"github.com/prometheus/client_golang/prometheus"
 )
 
@@ -263,7 +264,7 @@ func main() {
 	flag.Parse()
 
 	if *showVersion {
-		fmt.Println(GetVersion())
+		fmt.Println(version.GetVersion())
 		os.Exit(0)
 	}
 
@@ -709,18 +710,4 @@ func (e *Exporter) collectDeploymentMetrics(ch chan<- prometheus.Metric) error {
 	deploymentTaskGroupUnhealthyAllocs.Collect(ch)
 
 	return nil
-}
-
-// Version is the version, duh
-var Version string
-
-// Date is the build date
-var Date string
-
-// Commit is the commit that was built
-var Commit string
-
-// GetVersion returns the binary version
-func GetVersion() string {
-	return fmt.Sprintf("nomad-exporter Version: %s Commit: %s Date: %s", Version, Commit, Date)
 }
